12월13일 작업
ㄱ

+++ ajax/proc_ajax.php
@@ -248,6 +248,7 @@ switch($actty) {
 	   $buy_mem_idx = get_any("odr", "mem_idx" , "odr_idx = $odr_idx");
 	   $sell_mem_idx = get_any("odr", "sell_mem_idx" , "odr_idx = $odr_idx");
 	   $part_idx = get_any("odr_det", "part_idx" , "odr_det_idx = $actidx");
+
 	   update_val("odr_history","confirm_yn","Y", "odr_history_idx", $actidx);
 
 	   $sql = "insert into odr_history set 
@@ -272,14 +273,18 @@ switch($actty) {
 		   update_val("odr","complete_yn","Y", "odr_idx", $odr_idx);	   
 		}
 
-		//주문이 하나도 없을 시 part 삭제(조건절 : 나의 현재 주문 제외) 2016-12-09 박정권
-		$odr_cnt_check = QRY_CNT("odr_det","and part_idx ='".$part_idx."' and odr_idx <> ".$odr_idx." and (odr_status <> 0 and odr_status <> 99)") ;
+		//판매자가 취소할때만 주문이 한 건도 없을때 파트 삭제 2016-12-12 박정권
+		if ($sell_mem_idx == $_SESSION["MEM_IDX"])
+		{
+			$odr_cnt_check = QRY_CNT("odr_det","and part_idx ='".$part_idx."' and odr_idx <> ".$odr_idx." and (odr_status <> 0 and odr_status <> 99)") ;
+			
+			if ($odr_cnt_check == "0")
+			{	
+				$sql = "delete from part where part_idx ='".$part_idx."' ";				
+				$result=mysql_query($sql,$conn) or die ("SQL ERROR : ".mysql_error());
+			}
+		}		
 		
-		if ($odr_cnt_check == "0")
-		{	
-			$sql = "delete from part where part_idx ='".$part_idx."' ";				
-			$result=mysql_query($sql,$conn) or die ("SQL ERROR : ".mysql_error());
-		}
 		break;
 	//-- 종료 및 입금 처리 : 구매자가 수령했다는 내용을 판매자가 확인 하고, 완료 버튼을 눌렀을 때 ----------------------------------------
    case "CF2":
